+++ htdocs/js/playlists.js
@@ -162,7 +162,7 @@ function populatePlaylistSelect(obj, playlistSelectId, selectedPlaylist) {
 
 /**
  * Appends entries to a playlist
- * @param {string} type one of song, stream, dir, search
+ * @param {string} type one of song, stream, dir, search, album, disc, searchdir
  * @param {Array} uris uris to add
  * @param {string} plist playlist to append the uri
  * @param {Function} callback response handling callback
@@ -170,6 +170,12 @@ function populatePlaylistSelect(obj, playlistSelectId, selectedPlaylist) {
  */
 function appendPlaylist(type, uris, plist, callback) {
     switch(type) {
+        case 'searchdir':
+            sendAPI("MYMPD_API_PLAYLIST_CONTENT_APPEND_SEARCH", {
+                "expression": createBaseSearchExpression(uris[0], uris[1]),
+                "plist": plist
+            }, callback, true);
+            break;
         case 'song':
         case 'stream':
         case 'dir':
@@ -204,7 +210,7 @@ function appendPlaylist(type, uris, plist, callback) {
 
 /**
  * Inserts entries into a playlist
- * @param {string} type one of song, stream, dir, search
+ * @param {string} type one of song, stream, dir, search, album, disc, searchdir
  * @param {Array} uris uris to add
  * @param {string} plist playlist to insert the uri
  * @param {number} to position to insert
@@ -213,6 +219,12 @@ function appendPlaylist(type, uris, plist, callback) {
  */
 function insertPlaylist(type, uris, plist, to, callback) {
     switch(type) {
+        case 'searchdir':
+            sendAPI("MYMPD_API_PLAYLIST_CONTENT_INSERT_SEARCH", {
+                "expression": createBaseSearchExpression(uris[0], uris[1]),
+                "plist": plist
+            }, callback, true);
+            break;
         case 'song':
         case 'stream':
         case 'dir':
@@ -251,7 +263,7 @@ function insertPlaylist(type, uris, plist, to, callback) {
 
 /**
  * Replaces a playlist
- * @param {string} type one of song, stream, dir, search
+ * @param {string} type one of song, stream, dir, search, album, disc, searchdir
  * @param {Array} uris uris to add
  * @param {string} plist playlist to replace
  * @param {Function} callback response handling callback
@@ -259,6 +271,12 @@ function insertPlaylist(type, uris, plist, to, callback) {
  */
 function replacePlaylist(type, uris, plist, callback) {
     switch(type) {
+        case 'searchdir':
+            sendAPI("MYMPD_API_PLAYLIST_CONTENT_REPLACE_SEARCH", {
+                "expression": createBaseSearchExpression(uris[0], uris[1]),
+                "plist": plist
+            }, callback, true);
+            break;
         case 'song':
         case 'stream':
         case 'dir':